Provide fping addresses via pipe, instead of an argument list.
#exec_arguments defaults to a no-op, but #exec_input does not.  This
means that exec_callbacks modules that do no use #exec_input need to
define a no-op themselves, as the default will still try to write data
to stdin.  (This may be addressed in the monitor at some point, it is a
bit of a gotcha.)

Rather than define an empty #exec_input method, make a functional one,
and remove #exec_arguments.  This will fare better if the input list is
excessively long anyway, and it fixes Error::EPIPE.